+++ packages/authhero/test/authentication-flows/universal.spec.ts
@@ -44,11 +44,20 @@ describe("universal", () => {
     const oauthClient = testClient(oauthApp, env);
 
     const session = await env.data.sessions.create("tenantId", {
-      session_id: "sessionId",
+      id: "sessionId",
       user_id: "email|userId",
       client_id: "clientId",
       expires_at: new Date(Date.now() + 1000).toISOString(),
       used_at: new Date().toISOString(),
+      device: {
+        last_ip: "",
+        initial_ip: "",
+        last_user_agent: "",
+        initial_user_agent: "",
+        initial_asn: "",
+        last_asn: "",
+      },
+      clients: [],
     });
 
     const response = await oauthClient.authorize.$get(
@@ -63,7 +72,7 @@ describe("universal", () => {
       },
       {
         headers: {
-          cookie: `tenantId-auth-token=${session.session_id}`,
+          cookie: `tenantId-auth-token=${session.id}`,
           origin: "https://example.com",
         },
       },