+++ packages/kysely/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @authhero/kysely-adapter
 
+## 6.0.0
+
+### Minor Changes
+
+- Update session entity
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.42.0
+
 ## 5.1.0
 
 ### Minor Changes